Initial plant growth system
Plants grow in a life cycle system; you are able to choose how many stages a plant has and what those stages look like. You are also able to choose a plan for growing that plan based on how many stages it has (by creating a GrowthPlan asset in the editor and assigning it to the GrowingPlant object).