Implemented tool scripts
Tools can now be used. Example is with watering can in scene Farm; simply pick it up with the left controller, water the seed in the pot with left click. More needs to be done to make sure either controller can pick it up and use it (add a few more variables).

To create a new tool, simply derive a new script from the Tool script or apply a base tool script to an XR object. Fill in the exposed public variables (make sure it has an empty object with a SphereCollider trigger on it WITH a proper tag on that collider; see WateringCup prefab). Select the button you want to use under use action.